age mixing implemented in formation hazard function and cumulative hazard updated
This file gives correct distributions of formation events, with age
mixing implemented, but only works correctly if mortality and
conception/birth are switched off and if initial number of people and
number of people is the same. In a future version of this file, the
updating of the cumulative hazard will need to be modified so that it
also works correctly for populations with demographic dynamics.